services/stt-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
vosk==0.3.45
numpy==1.24.3
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
kafka-python==2.0.2
avro-python3==1.10.2
